cleanup, travis integration and dependency update

+++ app/models/folder.rb
@@ -1,13 +1,14 @@
+# frozen_string_literal: true
 module Folder
-  MAIL_ROOT_FOLDER = "/srv/vmail"
-  MAIL_USER = "vmail"
+  MAIL_ROOT_FOLDER = '/srv/vmail'.freeze
+  MAIL_USER = 'vmail'.freeze
   MAIL_GROUP = MAIL_USER
-  STORAGE_CLASS = "mdbox/storage" # Possible values "mdbox", "Maildir", "mbox"
-  
+  STORAGE_CLASS = 'mdbox/storage'.freeze # Possible values "mdbox", "Maildir", "mbox"
+
   def move_folder
-    if ENV["RACK_ENV"] == "production"
+    if ENV['RACK_ENV'] == 'production'
       if File.directory?(get_folder(name_change.last))
-        error!("A folder with this name already exists.", 400)
+        error!('A folder with this name already exists.', 400)
       else
         FileUtils.mv get_folder(name_change.first), get_folder(name_change.last)
       end
@@ -15,16 +16,14 @@ module Folder
   end
 
   def remove_folder
-    if ENV["RACK_ENV"] == "production"
-      FileUtils.rm_rf get_folder
-    end
+    FileUtils.rm_rf get_folder if ENV['RACK_ENV'] == 'production'
   end
-  
+
   def change_permissions(folder_array)
-    current_folder = ""
+    current_folder = ''
     folder_array.each do |folder|
-      current_folder += folder + "/"
-      FileUtils.chown MAIL_USER, MAIL_GROUP, MAIL_ROOT_FOLDER + "/" + current_folder
+      current_folder += folder + '/'
+      FileUtils.chown MAIL_USER, MAIL_GROUP, MAIL_ROOT_FOLDER + '/' + current_folder
     end
   end
 end
